pyproject.toml
[project]
name = "forum-fastapi"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiosqlite>=0.22.1",
    "email-validator>=2.3.0",
    "fastapi-swagger-ui-theme>=0.1.0",
    "fastapi[standard]>=0.128.0",
    "pwdlib[argon2]>=0.3.0",
    "pydantic-settings>=2.12.0",
    "pyjwt>=2.11.0",
    "sqlalchemy>=2.0.46",
]

[tool.ruff.lint]
select = ["ALL"]
ignore = [
    "ANN201",  
    "COM812",  # missing-trailing-comma
    "CPY001",  # Missing copyright notice at top of file
    "D100",    # Missing docstring in public module
    "D101",    # Missing docstring in public class
    "D102",
    "D103",    # Missing docstring in public function
    "D104",    # Missing docstring in public package
    "D107",
    "D203",    # blank line required before class docstring
    "D211",    # no-blank-line-before-class
    "D213",    # multi-line-summary-second-line
    "EM101",   # raw-string-in-exception
    "FIX002",  # line-contains-todo
    "ISC001",  # Conflicts with formatter ruff
    "INP001",
    "PLR0904", # Too many public methods (... > 20)
    "TC001",
    "TC002",
    "TD002",   # Missing author in TODO `# TODO(<author_name>): ...`
    "TD003",   # missing-todo-link
    "TRY003",  # raise-vanilla-args
]

[tool.ruff]
line-length = 120

